fix - 一些杂七杂八的优化和修复，还有额外的测试脚本
https://github.com/nuskey8/Lua-CSharp/pull/99/files
https://github.com/nuskey8/Lua-CSharp/pull/102/files
https://github.com/nuskey8/Lua-CSharp/pull/101/files
https://github.com/nuskey8/Lua-CSharp/pull/107

+++ src/Lua/CodeAnalysis/Compilation/Descriptions.cs
@@ -1,11 +1,10 @@
-using Lua.Internal;
 using Lua.Runtime;
 
 namespace Lua.CodeAnalysis.Compilation
 {
     public readonly record struct LocalVariableDescription
     {
-        public required byte RegisterIndex { get; init; }
+        public required ushort RegisterIndex { get; init; }
     }
 
     public readonly record struct FunctionDescription
@@ -19,7 +18,7 @@ namespace Lua.CodeAnalysis.Compilation
     {
         public required ReadOnlyMemory<char> Name { get; init; }
         public required int Index { get; init; }
-        public required byte RegisterIndex { get; init; }
+        public required ushort RegisterIndex { get; init; }
     }
 
     public readonly record struct GotoDescription
