+++ src/Lua/Runtime/Chunk.cs
@@ -14,7 +14,7 @@ public sealed class Chunk
     public required Chunk[] Functions { get; init; }
     public required int ParameterCount { get; init; }
     
-    public required byte MaxStackPosition { get; init; }
+    public required ushort MaxStackPosition { get; init; }
 
     Chunk? rootCache;
